rename: featuers/chatbot/components index 관리(#206)
