--- src/features/auth/components/index.ts
@@ -1,3 +1,0 @@
-export { AgreementItem } from './AgreementItem';
-export { EditProfileDialog } from './EditProfileDialog';
-export { SignupForm } from './SignupForm';
