+++ src/features/auth/index.ts
@@ -1,3 +1,4 @@
+// components
 export { AgreementItem } from './components/AgreementItem';
 export { SignupForm } from './components/SignupForm';
 export { EditProfileDialog } from './components/EditProfileDialog';
