+++ src/features/chatbot/components/Header.tsx
@@ -14,4 +14,4 @@ const Header = () => {
   );
 };
 
-export default Header;
+export { Header };
