+++ src/features/chatbot/components/WizAvatar.tsx
@@ -6,4 +6,4 @@ function WizAvatar() {
   );
 }
 
-export default WizAvatar;
+export { WizAvatar };
