+++ src/features/chatbot/index.ts
@@ -1,0 +1,6 @@
+export { ActionProvider } from './components/ActionProvider';
+export { Header } from './components/Header';
+export { MessageParser } from './components/MessageParser';
+export { WidgetButton } from './components/WidgetButton';
+export type { WidgetButtonProps } from './components/WidgetButton';
+export { WizAvatar } from './components/WizAvatar';
